Introduce ApiAssetRepository which extends base DapperAssetRepository
This avoids issues with methods on DapperAssetRepository requiring
context and causing lifecycle issues
