+++ DLCS.Repository/Assets/AssetRepository.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using DLCS.Core.Types;
 using DLCS.Model.Assets;
 
@@ -17,15 +16,9 @@ namespace DLCS.Repository.Assets
             this.dlcsContext = dlcsContext;
         }
 
-        public Task<Asset?> GetAsset(string id)
-        {
-            return GetAsset(id, false);
-        }
+        public Task<Asset?> GetAsset(string id) => GetAsset(id, false);
 
-        public Task<Asset?> GetAsset(AssetId id)
-        {
-            return GetAsset(id, false);
-        }
+        public Task<Asset?> GetAsset(AssetId id) => GetAsset(id, false);
 
         public async Task<Asset?> GetAsset(string id, bool noCache)
             => await dlcsContext.Images.FindAsync(id);
@@ -35,16 +28,5 @@ namespace DLCS.Repository.Assets
 
         public async Task<ImageLocation> GetImageLocation(AssetId assetId)
             => await dlcsContext.ImageLocations.FindAsync(assetId.ToString());
-        
-        public Task<PageOfAssets?> GetPageOfAssets(int customerId, int spaceId, int page, int pageSize, string orderBy, bool descending,
-            CancellationToken cancellationToken)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public Task Save(Asset asset, CancellationToken cancellationToken)
-        {
-            throw new System.NotImplementedException();
-        }
     }
 }
